Changed variables from float32 to float64

+++ models/calc.go
@@ -2,10 +2,10 @@ package models
 
 import "math"
 
-func FahrenheitToCelsius(in float32) float32 {
-	return float32(math.Round((float64(in)-32)*5/9*10) / 10)
+func FahrenheitToCelsius(in float64) float64 {
+	return math.Round((in-32)*5/9*10) / 10
 }
 
-func MphToKmh(in float32) float32 {
-	return float32(math.Round((float64(in) * 1.60934) * 10 / 10))
+func MphToKmh(in float64) float64 {
+	return math.Round((in * 1.60934) * 10 / 10)
 }
