+++ models/weatherData.go
@@ -1,29 +1,29 @@
 package models
 
 type WeatherData struct {
-	TempF          float32 `form:"tempf"`
-	TempC          float32
-	Humidity       float32 `form:"humidity"`
-	DewptF         float32 `form:"dewptf"`
-	DewptC         float32
-	WindChillF     float32 `form:"windchillf"`
-	WindChillC     float32
+	TempF          float64 `form:"tempf"`
+	TempC          float64
+	Humidity       float64 `form:"humidity"`
+	DewptF         float64 `form:"dewptf"`
+	DewptC         float64
+	WindChillF     float64 `form:"windchillf"`
+	WindChillC     float64
 	WindDir        uint8   `form:"winddir"`
-	WindSpeedMPH   float32 `form:"windspeedmph"`
-	WindSpeedKmh   float32
-	WindGustMPH    float32 `form:"windgustmph"`
-	WindGustKmh    float32
-	Rainin         float32 `form:"rainin"`
-	DailyRainin    float32 `form:"dailyrainin"`
-	WeeklyRainin   float32 `form:"weeklyrainin"`
-	MonthlyRainin  float32 `form:"monthlyrainin"`
-	YearlyRainin   float32 `form:"yearlyrainin"`
-	SolarRadiation float32 `form:"solarradiation"`
+	WindSpeedMPH   float64 `form:"windspeedmph"`
+	WindSpeedKmh   float64
+	WindGustMPH    float64 `form:"windgustmph"`
+	WindGustKmh    float64
+	Rainin         float64 `form:"rainin"`
+	DailyRainin    float64 `form:"dailyrainin"`
+	WeeklyRainin   float64 `form:"weeklyrainin"`
+	MonthlyRainin  float64 `form:"monthlyrainin"`
+	YearlyRainin   float64 `form:"yearlyrainin"`
+	SolarRadiation float64 `form:"solarradiation"`
 	UV             uint8   `form:"UV"`
-	IndoorTempF    float32 `form:"indoortempf"`
-	IndoorTempC    float32
+	IndoorTempF    float64 `form:"indoortempf"`
+	IndoorTempC    float64
 	IndoorHumidity uint8   `form:"indoorhumidity"`
-	BaroMin        float32 `form:"baromin"`
+	BaroMin        float64 `form:"baromin"`
 	DateUTC        string  `form:"dateutc"`
 	SoftwareType   string  `form:"softwaretype"`
 	Action         string  `form:"action"`
